Generalize error response
