--- lib/pluggy_elixir/error/unauthorized.ex
@@ -1,11 +1,0 @@
-defmodule PluggyElixir.Error.Unauthorized do
-  @moduledoc """
-  Error for invalid credentials.
-  """
-
-  defstruct [:message]
-
-  @type t :: %__MODULE__{
-          message: binary()
-        }
-end
