+++ test/pluggy_elixir/http_client_test.exs
@@ -1,11 +1,11 @@
 defmodule PluggyElixir.HttpClientTest do
-  use PluggyElixir.Case, async: true
+  use PluggyElixir.Case, async: false
 
   alias PluggyElixir.Auth
   alias PluggyElixir.Auth.Guard
-  alias PluggyElixir.Error.Unauthorized
   alias PluggyElixir.HttpAdapter.Response
   alias PluggyElixir.HttpClient
+  alias PluggyElixir.HttpClient.Error
 
   describe "[ authentication flow ]" do
     test "create api token when there inst and make get request", %{bypass: bypass} do
@@ -42,7 +42,7 @@ defmodule PluggyElixir.HttpClientTest do
 
       response = HttpClient.get(url)
 
-      assert {:error, %Unauthorized{} = error} = response
+      assert {:error, %Error{} = error} = response
 
       pid = inspect(self())
 
@@ -112,7 +112,7 @@ defmodule PluggyElixir.HttpClientTest do
 
       response = HttpClient.get(url)
 
-      assert {:error, %Unauthorized{}} = response
+      assert {:error, %Error{}} = response
     end
   end
